Report fitness as long rather than double

+++ com-mmxlabs-optimiser/optimiser-lso/src/main/java/com/mmxlabs/optimiser/lso/impl/LinearSimulatedAnnealingFitnessEvaluator.java
@@ -210,7 +210,7 @@ public final class LinearSimulatedAnnealingFitnessEvaluator<T> implements
 	 * 
 	 * @return
 	 */
-	public double getBestFitness() {
+	public long getBestFitness() {
 		return bestFitness;
 	}
